OM-1384 Modify db_t methods to not raise error when given empty key
Previously this would call underlying #t method, which would raise an error
if given an empty string. But since the keys are dynamic, it seems more likely
to occur, so instead just return default if provided, otherwise return the key
(nil or empty string). Find / fix callers to not call db_t with empty string
in another commit