OM-1423 Bug fix: census data types with parens dont match profile config
School profile config tables can filter out data based on data type and
breakdown.

When querying for Characteristics(census) data for the school profile,
we filter out data that doesn't match the criteria set in the profile
config.

In the conditional that matches CensusDataTypes with
CategoryData#response_keys, we were doing a String#matches(/some_string/)
but some_string sometimes contained parens. Since we're passing
String#matches a regular expression, we need to Regexp#escape the string
that is passed in. This fixes the issue when the input string contains
parenthesis.