AT-961 Don't create group_comparisons for empty data.
Some data sets were creating bar chart collections and
bar charts that had no underlying bar chart bars.
This caused a slice error on an empty array.

This commit adds a check so that bar chart bars have to be present
and also asserts that we don't slice empty arrays.