JT-78 Route to register new emails, be able to subscribe to lists via email
A new route needed to be added that will register a new email in list member
if it doesnt already exist. Will not respond with any errors if email is taken.

SubscriptionConcerns now has classes that know how to sign someone up via email
rather than depending on current_user in session. Email param takes precedence if
available, but falls back to current_user

Javascript for modals was modified so that modal data (such as email) can be
returned from resolved/rejected modals, so that it can turn be given to
JS subscription methods