JT-37 Remove provisional check when validating that user can log in with email
When user types in email to log in, an ajax call validates the email and
possibly checks messages. The server will no longer respond with error
if the email belongs to a provisional account.

Also renamed the route and method that performs above check to represent
what it actually does now