JT-138 Fix has_subscription? in User and use it in safely_add_subscription!
The has_subscription? method wasn't defaulting state and school ID to
CA and 0, but it should since those are the default values that get
set since state and school ID are not nullable on list_active.

Safely_add_subscription! was using is_signedup? which is redundant code
and doesnt take school into consideration, so changed it to use the fixed
has_subscription? method