JT-106 Add signin link in parsley error, and make it work correctly
There is a message that gets shown to the user if they type an existing
email into the join form. We want a link to be in that message so the
user can click it to see the signin form instead.

Adding the link is trivial, but it only worked "only once" in that if you
clicked it, the signin form would be shown, but if you clicked back to
the join form using the tab, and then clicked the link in the message,
nothing would happen.