JT-129 Submit review via ajax after join modal resolves.
Previously we would save a review via ajax using rails' built-in
remote:true on forms, and then show the Join modal. We instead want
to use the Join modal to get a user registered or signed in, and then
post their review manually via Ajax. If the user is a new user, don't
refresh the page, and instead show the form that asks the user to tell
us their affiliation with the school (user type), then refresh the page.
Always refresh the page if the user already existed.