JT-379 Refactor reset_password route handling code to be more generic
Previously, we had a route specific to the Reset Password flow that would
authenticate a user using a token, and then redirect them to a hardcoded
password page url. Instead, make a brand new route and action for allowing
a redirect param to be set. Authenticate the user and then redirect there.

Followup refactoring will need to be made to make this code be shared with
another #verify_email action in SigninController